Bug 170772 - Unable to launch RCP application
Changed logic to always parse osgi.bundles (whether we generate it or it comes from a template).  By doing this, we can run the extra logic to check to see if org.eclipse.update.configurator is included in the osgi.bundles.  If it is not, we need to list each bundle location in the osgi.bundles.

Also added a two line fix to make sure if config.ini does not exist in target platform, that we launch with a generated config.ini.