Modified the fix for Bug 170772 (Unable to launch RCP application).
In the original fix, we parsed/recomputed the osgi.bundles if osgi.bundles was present (either from a template or from the target's config.ini).

After talking to Wassim, I corrected the logic to only parse/compute the osgi.bundles if the value is present in the target's config.ini (ie. we are generating a config.ini).

I factored out the computation logic into a function (computeOSGiBundles).  This is now called after reading the target's config.ini if the file contained an osgi.bundles tag or if no osgi.bundles tag was specified in the target's config.ini or user defined template.