Bug 210131 - Extension Registry enhancement - include disabled plug-ins
Ran into a problem closing a project that had an external equivalent.  This was caused because we need to refresh the extension registry with the external plug-in.  Since the registry ties the extension point to only one plug-in, when we remove all workspace codes the extension point is tied to the old workspace copy.  Without refreshing, the registry will thing the extension point no longer exists.