Bug 214542 - Error importing source plugins stored in folders
Error happened as the source plug-in created a scenario where we tried to copy contents of a folder to itself.  Before the copy happened, we deleted the destination (which also deleted the source), which caused us to try to copy something that did not exist.  Therefore, the fix was to simply not do anything if we were going to copy a folder to itself.